requirements.txt
Django==3.2.12
gunicorn==20.1.0
djangorestframework==3.15.1